refactor: remove custom types and consolidate packages
- Remove visitedRefs type, replace with built-in map[string]struct{}
- Move path utilities from utils to diff package
- Remove lint package
