--- lint/doc.go
@@ -1,4 +1,0 @@
-/*
-Package lint provides a lint function for OpenAPI Spec 3.
-*/
-package lint
